# Fix z-fighting and performance issues in 2D infinite scroll mode