--- a/packages/engine/Source/Scene/Scene.js
+++ b/packages/engine/Source/Scene/Scene.js
@@ -3310,8 +3310,9 @@ function execute2DViewportCommands(scene, passState) {
   }
 
   camera._setTransform(transform);
-  Cartesian3.clone(position, camera.position);
-  camera.frustum = frustum.clone();
+  camera.position.x = position.x;
+  camera.position.y = position.y;
+  camera.frustum = frustum;
   passState.viewport = originalViewport;
 }